fix(news-hub): address PR #598 R24 — inline useHydrated for §17, WHY comment, useWaitForMarketNewsCards extraction, content key, refresh-flag rationale
